- XcspSettings now match Configuration API - clean up code
git-svn-id: svn://svn.code.sf.net/p/choco/code/trunk@1615 5779b9a9-2447-0410-887e-bb98d57c6de4

+++ choco-tools/choco-parsers/src/main/java/parser/flatzinc/ast/expression/EIdArray.java
@@ -59,12 +59,12 @@ public final class EIdArray extends Expression{
 
     @Override
     public String toString() {
-        return name+"["+index.toString()+"]";
+        return name+ '[' +index.toString()+ ']';
     }
 
     @Override
     public int intValue() {
-        return ((Integer)object).intValue();
+        return (Integer) object;
     }
 
     @Override
